update configuration to belong to devices, remove json checks
