+++ src/main/java/org/beanpod/switchboard/dto/DeviceDto.java
@@ -1,5 +1,8 @@
 package org.beanpod.switchboard.dto;
 
+import java.time.OffsetDateTime;
+import javax.persistence.Column;
+import javax.persistence.Lob;
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,4 +20,8 @@ public class DeviceDto {
   @NotNull private String privateIpAddress;
   @NotNull private String displayName;
   private String status;
+  @Lob private byte[] configurationInstance;
+
+  @Column(name = "configuration_last_modified")
+  private OffsetDateTime configurationLastModified;
 }
