+++ src/main/java/org/beanpod/switchboard/dto/EncoderDto.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.Date;
 import java.util.Set;
-import javax.persistence.Column;
-import javax.persistence.Lob;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import javax.validation.constraints.NotNull;
@@ -34,12 +32,4 @@ public class EncoderDto {
 
   @JsonIgnoreProperties("encoder")
   private Set<OutputChannelDto> output;
-
-  @Lob private byte[] configurationSchema;
-  @Lob private byte[] configurationInstance;
-
-  @Column(name = "configuration_last_modified")
-  @Temporal(TemporalType.TIMESTAMP)
-  @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-  private Date configurationLastModified;
 }
