+++ src/main/java/org/beanpod/switchboard/entity/DeviceEntity.java
@@ -1,10 +1,12 @@
 package org.beanpod.switchboard.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.time.OffsetDateTime;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.Lob;
 import javax.persistence.OneToOne;
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -52,4 +54,11 @@ public class DeviceEntity {
       mappedBy = "device",
       cascade = {CascadeType.REMOVE})
   private EncoderEntity encoderEntity;
+
+  @Lob
+  @Column(name = "configuration_instance")
+  private byte[] configurationInstance;
+
+  @Column(name = "configuration_last_modified")
+  private OffsetDateTime configurationLastModified;
 }
