+++ src/main/java/org/beanpod/switchboard/service/EncoderServiceImpl.java
@@ -2,18 +2,14 @@ package org.beanpod.switchboard.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.beanpod.switchboard.dao.EncoderDaoImpl;
 import org.beanpod.switchboard.dao.StreamDaoImpl;
 import org.beanpod.switchboard.dto.EncoderDto;
 import org.beanpod.switchboard.dto.StreamDto;
 import org.beanpod.switchboard.exceptions.ExceptionType;
-import org.beanpod.switchboard.exceptions.ExceptionType.InvalidJsonException;
 import org.beanpod.switchboard.util.DateUtil;
-import org.beanpod.switchboard.util.JsonUtil;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
 @Component
 @RequiredArgsConstructor
@@ -37,18 +33,5 @@ public class EncoderServiceImpl implements EncoderService {
     log.debug("Updated encoder {} last communication date", encoderSerialNumber);
 
     return streamDao.getEncoderStreams(encoderSerialNumber);
-  }
-
-  @SneakyThrows
-  @Override
-  public void uploadJson(MultipartFile schema, MultipartFile instance) {
-    boolean isSchemaValid = JsonUtil.isJsonValid(new String(schema.getBytes()));
-    boolean isInstanceValid = JsonUtil.isJsonValid(new String(instance.getBytes()));
-    if (!isSchemaValid) {
-      throw new InvalidJsonException("configuration schema.");
-    }
-    if (!isInstanceValid) {
-      throw new InvalidJsonException("Instance schema.");
-    }
   }
 }
